Implement typed maps (#36)
This PR implements typed maps in martian-rust. In other words, HashMaps that you use in the inputs and outputs of stages, will no longer be map in the generated MRO, they will be map<type>. This allows martian to do better static analysis of the Cellranger pipeline, and also enables Rust stages to pass around HashMaps holding MartianFileTypes with Martian properly tracking the files within (previously it was basically required to workaround this e.g. by passing around a Vector of key-value pairs for example.)

The main drawback of this implementation is that HashMaps are now required to have keys that implement Hash + Display + Eq and values that implement MartianPrimaryType. This is the case for most of the uses of HashMaps in the Cellranger pipeline, AFAIK there were just two cases where this was not true. For cases where your HashMap doesn't meet those bounds, it is required to use #[mro_type = "map"] to force the map to be untyped. AFAIK this drawback will continue to be the case until Specialization is in stable Rust, allowing there to be one implementation of MartianBlanketType for any HashMaps meeting the trait bounds and another for ones that do not.