Mark error pathways as #[cold].
This is a hint to the compiler that it should optimize those paths for
size more than speed, and probably move them out of the main body of
code so that they don't slow down loading on the happy path.  We don't
really care if error reporting pathways are slower, anyway.