Switch from `chrono` to `time` crate.
`chrono` appears to be unmaintained, with no updates in over a year
and active security advisories open against it. Other crates (such
as for example junit-report) have started migrating to the `time`
crate, so this is following suit.

Bumping the version on martian and martian-derive due to changes here
and in earlier commits.