Various fixes and cleanups (including breaking changes) (#133)
* Remove some unnecssary trait bounds.

We don't need stage inputs to be serializable, and we don't need stage
outputs to be deserializable.  Since implementations of those traits can
add a lot of generated code to downstream crates, it's best to leave
those constraints out.

Also remove some unnecessary `Clone` trait bounds, and allow
LazyFileTypeIO::read_all to read directly into containers other than
vectors (anything that implements FromIterator, e.g. HashSet).

Move a bit more code into places where it won't be duplicated for each
type (in this case MartianFileType) by adding a from_path method to the
trait, which takes a single &Path and deals with splitting it, rather
than requiring the caller to do so.

* Switch some Strings to strs.

Also, more work to reduce the amount of code being generated by
monomorphization.

* Close file before updating journal.

NFS makes no guarantee that bytes are actually flushed to the server
until the file is closed, so updating the journal before the file is
closed could result in mrp seeing an incomplete write.

* Switch from `chrono` to `time` crate.

`chrono` appears to be unmaintained, with no updates in over a year
and active security advisories open against it. Other crates (such
as for example junit-report) have started migrating to the `time`
crate, so this is following suit.

Bumping the version on martian and martian-derive due to changes here
and in earlier commits.

Co-authored-by: sreenathkrishnan <sreenath.krishnan@10xgenomics.com>