Implement a helper trait for creating temporary martian files. (#347)
Adds the `MartianTempFile` trait, providing a `tempfile()` method that creates a temporary file that complies with the naming expectations of the file type.  The `TypedTempFile<F>` returned by the method owns a temporary file and thus manages the tempfile lifetime, and implements `Deref` and `AsRef` for the underlying file type.